Add end-to-end keccak test for using keccak accelerator in the guest
* Add control IDs and control roots for keccak (we may want to move to a single control root for the lift predicates)
* Update dependency on risc0 repo

Still todo:

* more cleanup
* CUDA
