+++ .gitignore
@@ -7,6 +7,7 @@
 .idea/
 bazel-*
 Cargo.lock
+!**/methods/**/Cargo.lock
 dist/
 rust-project.json
 target/
