+++ Cargo.toml
@@ -4,43 +4,48 @@ members = ["zirgen/circuit/fib", "zirgen/circuit/keccak", "zirgen/circuit/keccak
 
 [workspace.dependencies.risc0-binfmt]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
+default-features = false
 
 [workspace.dependencies.risc0-build]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
+default-features = false
 
 [workspace.dependencies.risc0-core]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
 default-features = false
 
 [workspace.dependencies.risc0-zkp]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
 default-features = false
 
 [workspace.dependencies.risc0-zkvm]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
 default-features = false
 
 [workspace.dependencies.risc0-circuit-recursion]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
 default-features = false
 
 [workspace.dependencies.risc0-zkvm-platform]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
+default-features = false
 
 [workspace.dependencies.risc0-sys]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
+default-features = false
 
 [workspace.dependencies.risc0-build-kernel]
 git = "https://github.com/risc0/risc0.git"
-rev = "16fea37265699f75d0c1a52a552ffed229a96fed"
+rev = "5a2ad2b3bdd197123cf0fbfdf7aec88150105da3"
+default-features = false
 
 [workspace.dependencies.risc0-zirgen-dsl]
 path = "zirgen/dsl"
@@ -67,4 +72,3 @@ opt-level = 3
 
 [workspace.dependencies]
 sppark = "0.1.8"
-
