+++ rust-toolchain.toml
@@ -1,4 +1,5 @@
 [toolchain]
-channel = "stable"
+# mirror zkvm
+channel = "1.77"
 components = ["clippy", "rustfmt"]
 profile = "minimal"
