+++ zirgen/bootstrap/src/main.rs
@@ -94,6 +94,7 @@ const KECCAK_SYS_OUTPUTS: &[&str] = &[
     "rust_poly_fp_3.cpp",
     "rust_poly_fp_4.cpp",
 ];
+const KECCAK_ZKR_ZIP: &str = "keccak_zkr.zip";
 
 #[derive(Clone, Debug, ValueEnum)]
 enum Circuit {
@@ -265,10 +266,15 @@ impl Args {
     }
 
     fn keccak(&self) {
+        let bazel_bin = get_bazel_bin();
         let out = &self.output;
         let circuit = "keccak";
         let src_path = Path::new("zirgen/circuit/keccak");
         let sys_root = Path::new("zirgen/circuit/keccak-sys").to_path_buf();
+        let zkr_src_path = bazel_bin.join("zirgen/circuit/predicates");
+        let zkr_tgt_path = out
+            .clone()
+            .unwrap_or(Path::new("zirgen/circuit/keccak/src/prove").to_path_buf());
 
         copy_group(circuit, &src_path, out, ZIRGEN_RUST_OUTPUTS, "src", "");
         copy_group(
@@ -279,6 +285,7 @@ impl Args {
             "cxx",
             "",
         );
+        copy_file(&zkr_src_path, &zkr_tgt_path, KECCAK_ZKR_ZIP);
         cargo_fmt_circuit(circuit, &self.output, &None);
     }
