+++ zirgen/circuit/BUILD.bazel
@@ -10,6 +10,7 @@ filegroup(
         "//zirgen/circuit/fib",
         "//zirgen/circuit/keccak",
         "//zirgen/circuit/predicates:gen_predicates",
+        "//zirgen/circuit/predicates:keccak_zkr",
         "//zirgen/circuit/recursion",
         "//zirgen/circuit/rv32im/v1/edsl:rv32im",
         "//zirgen/dsl/examples/calculator",
