+++ zirgen/circuit/keccak/methods/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "risc0-circuit-keccak-test-methods"
+version = "0.1.0"
+edition = "2021"
+
+[build-dependencies]
+risc0-build = { workspace = true }
+
+[package.metadata.risc0]
+methods = ["guest"]
