+++ zirgen/circuit/keccak/methods/guest/src/bin/keccak.rs
@@ -1,0 +1,22 @@
+// Copyright 2024 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use risc0_zkvm::guest::env;
+use risc0_zkvm::sha::Digest;
+
+fn main() {
+    // Computes and proves the result of a given keccak input transcript
+    let (claim_digest, input, po2): (Digest, Vec<u32>, usize) = env::read();
+    risc0_circuit_keccak::prove(claim_digest, &input, po2).expect("Unable to prove with keccak");
+}
