+++ zirgen/circuit/keccak/src/control_id.rs
@@ -1,0 +1,45 @@
+// Copyright 2024 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Automatically generated by:
+//   cargo run -p risc0-circuit-keccak -F make_control_ids --bin make_control_ids > zirgen/circuit/keccak/src/control_id.rs
+
+use risc0_zkp::{core::digest::Digest, digest};
+
+pub const KECCAK_CONTROL_IDS: [Digest; 11] = [
+    digest!("4ad5ef0b1ba53b0be1fcfc35f05fd941a57c84054113d85086fd180e89241a0d"),
+    digest!("25d29b4e09e0273a6de5f04e0d6451676ab195365cbf1e480f1198314b75fb0d"),
+    digest!("e098062022286b5f0da3b9135e328b3c0e61e94bf553de29b9091749ca930a2e"),
+    digest!("c6732c3f9fdb7f54cfa43425552e4e67e7260201decdad0d40fdbd17b1f49849"),
+    digest!("5763c11059ffd554b5df5913270a78102664c56bf6b7f6416927eb15fa493c4f"),
+    digest!("059ce104bb0cd20e5f400a63613dbb087173145a31e00018b8afd25cbcef1529"),
+    digest!("939c474965736118d8cc9856dcb78d684fa53033481d6342e798ed0cc04d8612"),
+    digest!("99726d1fe4d40b367ceef95dbaa4b55cb4ce634ac12a861a0cc4a23a41442760"),
+    digest!("8823b91d58d891270f558c4ec6320d2517ba9a295db6af1425f9ed10f7892463"),
+    digest!("67af7649aa850f2e1f30e965f0e44442b5287f61ed6071646120c92c6c5e5618"),
+    digest!("828a1d287cd71e405269f55f70be311951e2ea1e9d4aa950e7fdf662d2d8d53f"),
+];
+pub const KECCAK_CONTROL_ROOTS: [Digest; 11] = [
+    digest!("6e720a2d96f9a433a3b4322f472720475aa291630176ba3cab19f22ed9594516"),
+    digest!("db056e02c945363d76884e634b75ae2fa5f68e2ae0924d71d51fde3325c8c404"),
+    digest!("b372a858430e6d2fa9eb6a680bec533080a9333ba20bdf1a78d5f52dfd364f2b"),
+    digest!("4711351ca72ea831ad17b311ccc2974f94dca8125257ef6b80ab4023e5214264"),
+    digest!("62d23003ed3063011a16be67db18c269de2da616178ddb5867731d2a8c4afe0b"),
+    digest!("fb21ed176e92c0547842cc68d9ec4633775e26128e86a3260058b8158660bc72"),
+    digest!("b0ed1c68e2b211008bcf0e6fe0d20d72b66af5080414702a277a84527f7c3a3f"),
+    digest!("438b5456e29b555d5acfb9431e637f5caee7ac49eefe26618ecf62650c790f3e"),
+    digest!("bf638a0947c93742cc0758347284f408c5c33c5aec907b6f19a1c665e073e244"),
+    digest!("cbfc486e5ec2426bf59c371b2417964a0b6682035ea598248af6521f1503151f"),
+    digest!("0ecba46b2fa56048fe790b6f02e5810d493f356225f05b4a134178337ffe8a34"),
+];
