+++ zirgen/circuit/predicates/BUILD.bazel
@@ -32,10 +32,12 @@ ZKRS = [
     "identity",
 ]
 
+KECCAK_ZKRS = [("keccak_lift_" + str(po2)) for po2 in range(14, 25)]
+
 build_circuit(
     name = "gen_predicates",
     srcs = ["gen_predicates.cpp"],
-    outs = [fn for zkr in ZKRS for fn in [
+    outs = [fn for zkr in (ZKRS + KECCAK_ZKRS) for fn in [
         zkr + ".zkr",
         zkr + ".zkr.dbg",
     ]],
@@ -70,4 +72,9 @@ sh_test(
         ":actual_hashes.txt",
         ":golden_hashes.txt",
     ],
+)
+
+pkg_zip(
+    name = "keccak_zkr",
+    srcs = [x + ".zkr" for x in KECCAK_ZKRS],
 )
